Rename module Ansible to AnsbileV1
Because we are implementing Ansbile version 2, we are renaming the module
as ansibleV1.
In future will be the possibility to switch Ansible version by configuration.
