+++ core/AnsibleV1Inv.py
